[MIG] mgmtsystem_nonconfomrmity_type: Migration to 13.0
