+++ mgmtsystem_nonconformity_type/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_nonconfrmity_type
+from . import test_nonconformity_type
